+++ lensless/apgd.py
@@ -208,7 +208,9 @@ class APGD(ReconstructionAlgorithm):
         )
 
     def reset(self):
-        if self._image_est is None:
+        if self._initial_est is not None:
+            self._image_est = self._initial_est
+        else:
             self._image_est = np.zeros(self._original_size, dtype=self._dtype)
 
     def _update(self, iter):